requirements.txt
pytest
pip
argparse
mark
requests
selenium
openpyxl
allure-pytest
pytest-xdist
